tiny fixes
- update ColorIndex to u16
- fix comment in cos_sin
- fix linear_gradient_palette items count

+++ src/lib.rs
@@ -2,7 +2,7 @@ use std::collections::BTreeMap;
 use std::fs::File;
 use std::io::{BufWriter, Write};
 
-pub type ColorIndex = u8;
+pub type ColorIndex = u16;
 pub type Palette = BTreeMap<ColorIndex, Color>;
 pub type RenderFunc = dyn Fn(RenderPoint) -> f32;
